feat: altera tipo de dados de estrela e nota para double nas classes AvaliacaoDto e Avaliacao

+++ src/main/java/com/example/vaideboa/Dtos/AvaliacaoDto.java
@@ -2,9 +2,9 @@ package com.example.vaideboa.Dtos;
 
 public class AvaliacaoDto {
   private final Long idReserva;  
-  private final int estrela;
+  private final double estrela;
   private final String mensagem;
-  public AvaliacaoDto(Long idReserva, int estrela, String mensagem) {
+  public AvaliacaoDto(Long idReserva, double estrela, String mensagem) {
     this.idReserva = idReserva;
     this.estrela = estrela;
     this.mensagem = mensagem;
@@ -12,7 +12,7 @@ public class AvaliacaoDto {
   public Long getIdReserva() {
     return idReserva;
   }
-  public int getEstrela() {
+  public double getEstrela() {
     return estrela;
   }
   public String getMensagem() {
